# perf(lint): use `Arc<T>` to share `PackageJson` among analyzer rules