## Summary

I instrumented Biome running in a large repository (~ 5,200 files) in my company, and observed a hotspot that `<ManifestServices as FromServices>::from_services` is relatively heavy. It is executed in every rule that queries `Manifest<T>`. 

Here is a screenshot of the hotspot in the instrumentation:

<img width="1404" alt="image" src="https://github.com/user-attachments/assets/5eedf105-b41b-4eee-8466-faea83745e00" />

In this pull request, `ManifestJson` will be wrapped in an `Arc<T>`. The heavy clone no longer happen.

## Test Plan

Existing tests should pass.

Also locally tested in the same repository to see the execution time:

Before

```
❯ time ./node_modules/.pnpm/@biomejs+cli-darwin-arm64@2.0.0-beta.2/node_modules/@biomejs/cli-darwin-arm64/biome check
Checked 5243 files in 7s. No fixes applied.

________________________________________________________
Executed in    8.16 secs    fish           external
   usr time   41.30 secs    0.19 millis   41.30 secs
   sys time   16.02 secs    1.23 millis   16.01 secs
```

After

```
❯ time ~/Projects/github.com/biomejs/biome/target/release-with-debug/biome check
Checked 5243 files in 6s. No fixes applied.

________________________________________________________
Executed in    7.36 secs    fish           external
   usr time   31.41 secs    0.24 millis   31.41 secs
   sys time   17.50 secs    2.30 millis   17.50 secs
```